⚡ refactor: #5 新方式を再びコンポーネント化
