+++ src/components/types.ts
@@ -4,8 +4,17 @@ export interface Panel {
   id: string;
   cells: CellType[][];
 }
-
 export interface CellTypeConfig {
   label: string;
   color: string;
+}
+
+export interface PanelPlacementModeType {
+  panel: Panel | null;
+  highlightedCell: { row: number; col: number } | null;
+}
+
+export interface PanelPlacementHistoryType {
+  panel: Panel | null;
+  highlightedCell: { row: number; col: number } | null;
 }
